Add files via upload
The single-line PHP script generates a random password string of a specified length.

Here is how it works:

The str_repeat() function repeats a string of characters. In this case, it repeats the string 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789' 8 times to ensure a good mix of uppercase and lowercase letters and digits in the password.
The str_shuffle() function shuffles the characters of the repeated string from step 1 to create a random sequence.
The substr() function returns a substring of the shuffled string, starting from the <a href="https://glowupclean.com/">first character</a>  and extending 8 characters.
The resulting password is a random 8-character string consisting of a mix of uppercase and lowercase letters and digits. You can adjust the length of the password by changing the second parameter in the substr() function.